refactor calendar to its own component
