+++ src/routes/app/events/+page.svelte
@@ -7,12 +7,12 @@
   import PageText from "$lib/components/PageText.svelte";
   import Contribute from "$lib/components/Contribute.svelte";
   import Date from "$lib/components/Date.svelte";
-  import { Calendar } from "bits-ui";
   import { parseAbsoluteToLocal } from "@internationalized/date";
-  import type { CalendarDate, DateValue } from "@internationalized/date";
+  import type { CalendarDate } from "@internationalized/date";
+  import Calendar from "$lib/components/Calendar.svelte";
 
   let { data }: PageProps = $props();
-  let value: CalendarDate | undefined = $state();
+  let selectedDate: CalendarDate | undefined = $state();
 
   function getDay(event: CalendarEventEnriched): string {
     return event.startDate.split("T")[0];
@@ -78,35 +78,14 @@
    * Filter events by selected calendar date
    */
   let filteredEvents = $derived.by(() => {
-    if (!value) {
+    if (!selectedDate) {
       return $eventsByDate;
     }
 
-    return $eventsByDate.filter((group) => group.date === value.toString());
-  });
-
-  /**
-   * Busy-ness indicator on highlighted dates
-   * Return a opacity value between 0 and 100.
-   * 0 = no events
-   * 100 = 5 or more events.
-   */
-  function getBusynessOpacity(date: DateValue): number {
-    if (!$eventsByDate) {
-      return 0;
-    }
-
-    const groupForDate = $eventsByDate.find(
-      (group) => group.date === date.toString(),
+    return $eventsByDate.filter(
+      (group) => group.date === selectedDate.toString(),
     );
-
-    if (!groupForDate) {
-      return 0;
-    }
-
-    const eventCount = Math.min(groupForDate.eventsList.length, 5);
-    return (eventCount / 5) * 100;
-  }
+  });
 </script>
 
 <CalendarSelector />
@@ -116,61 +95,18 @@
 {/if}
 
 {#if $eventsByDate && $eventsByDate.length > 0 && $calendar}
-  <Calendar.Root
+  <Calendar
     type="single"
+    bind:value={selectedDate}
+    busyness={$eventsByDate}
     minValue={$calendar.startDate
       ? parseAbsoluteToLocal($calendar.startDate)
       : undefined}
     maxValue={$calendar.endDate
       ? parseAbsoluteToLocal($calendar.endDate)
       : undefined}
-    bind:value
-  >
-    {#snippet children({ months, weekdays })}
-      <Calendar.Header class="flex flex-row">
-        <Calendar.PrevButton class="w-8 mr-2">←</Calendar.PrevButton>
-        <Calendar.Heading />
-        <Calendar.NextButton class="w-8 ml-2">→</Calendar.NextButton>
-      </Calendar.Header>
+  />
 
-      {#each months as month (month.value)}
-        <Calendar.Grid>
-          <Calendar.GridHead>
-            <Calendar.GridRow>
-              {#each weekdays as day, index (index)}
-                <Calendar.HeadCell>
-                  {day}
-                </Calendar.HeadCell>
-              {/each}
-            </Calendar.GridRow>
-          </Calendar.GridHead>
-          <Calendar.GridBody>
-            {#each month.weeks as weekDates, weekIndex (weekIndex)}
-              <Calendar.GridRow>
-                {#each weekDates as date (date)}
-                  <Calendar.Cell {date} month={month.value}>
-                    <Calendar.Day
-                      class={`border-black border
-                      rounded data-[outside-month]:pointer-events-none
-                      data-[outside-month]:text-gray-300
-                      data-[selected]:bg-black
-                      data-[selected]:text-white
-                      data-[disabled]:opacity-50
-                      data-[disabled]:border-0
-                      bg-physical/${getBusynessOpacity(date)}
-                      `}
-                    >
-                      {date.day}
-                    </Calendar.Day>
-                  </Calendar.Cell>
-                {/each}
-              </Calendar.GridRow>
-            {/each}
-          </Calendar.GridBody>
-        </Calendar.Grid>
-      {/each}
-    {/snippet}
-  </Calendar.Root>
   {#if filteredEvents.length > 0}
     {#each filteredEvents as group (group.date)}
       <div>
